# core, trie: new trie