This PR contains a rewrite of package trie with the following goals:

- reduced API surface: unexported node types, string funcs, other internals
- documented core internals
- documented API
- improved performance
- more memory reuse, especially when hashing the trie
- hashing is independent of the database