@fjl 2015-08-18
Preliminary performance report:

```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkInsertChain_empty_memdb-8              109218        64479         -40.96%
BenchmarkInsertChain_empty_diskdb-8             194874        136614        -29.90%
BenchmarkInsertChain_valueTx_memdb-8            240241        156505        -34.85%
BenchmarkInsertChain_valueTx_diskdb-8           412044        322234        -21.80%
BenchmarkInsertChain_valueTx_100kB_memdb-8      1203981       1010477       -16.07%
BenchmarkInsertChain_valueTx_100kB_diskdb-8     3067184       2969643       -3.18%
BenchmarkInsertChain_uncles_memdb-8             213715        170358        -20.29%
BenchmarkInsertChain_uncles_diskdb-8            301778        251242        -16.75%
BenchmarkInsertChain_ring200_memdb-8            144237467     81842282      -43.26%
BenchmarkInsertChain_ring200_diskdb-8           170420433     93856663      -44.93%
BenchmarkInsertChain_ring1000_memdb-8           194525358     116581937     -40.07%
BenchmarkInsertChain_ring1000_diskdb-8          219006749     127585191     -41.74%
BenchmarkCompactEncode                          135           61.4          -54.52%
BenchmarkCompactDecode                          87.8          87.2          -0.68%
BenchmarkCompactHexDecode                       116           113           -2.59%
BenchmarkDecodeCompact                          53.8          53.4          -0.74%
BenchmarkGet                                    628           289           -53.98%
BenchmarkGetDB                                  518           1471          +183.98%
BenchmarkUpdateBE                               2935          2002          -31.79%
BenchmarkUpdateLE                               3289          2409          -26.76%
BenchmarkHashBE                                 100965203     40623083      -59.77%
BenchmarkHashLE                                 117250940     43654768      -62.77%

benchmark                                       old allocs     new allocs     delta
BenchmarkInsertChain_empty_memdb-8              743            330            -55.59%
BenchmarkInsertChain_empty_diskdb-8             810            388            -52.10%
BenchmarkInsertChain_valueTx_memdb-8            1500           790            -47.33%
BenchmarkInsertChain_valueTx_diskdb-8           1616           890            -44.93%
BenchmarkInsertChain_valueTx_100kB_memdb-8      1504           794            -47.21%
BenchmarkInsertChain_valueTx_100kB_diskdb-8     1847           1029           -44.29%
BenchmarkInsertChain_uncles_memdb-8             1195           782            -34.56%
BenchmarkInsertChain_uncles_diskdb-8            1260           846            -32.86%
BenchmarkInsertChain_ring200_memdb-8            1182124        339278         -71.30%
BenchmarkInsertChain_ring200_diskdb-8           1220860        345026         -71.74%
BenchmarkInsertChain_ring1000_memdb-8           1600303        429111         -73.19%
BenchmarkInsertChain_ring1000_diskdb-8          1638553        435461         -73.42%
BenchmarkCompactEncode                          2              1              -50.00%
BenchmarkCompactDecode                          1              1              +0.00%
BenchmarkCompactHexDecode                       1              1              +0.00%
BenchmarkDecodeCompact                          1              1              +0.00%
BenchmarkGet                                    2              1              -50.00%
BenchmarkGetDB                                  2              6              +200.00%
BenchmarkUpdateBE                               14             8              -42.86%
BenchmarkUpdateLE                               13             8              -38.46%
BenchmarkHashBE                                 392188         106708         -72.79%
BenchmarkHashLE                                 473505         111063         -76.54%

benchmark                                       old bytes     new bytes     delta
BenchmarkInsertChain_empty_memdb-8              46288         19827         -57.17%
BenchmarkInsertChain_empty_diskdb-8             54900         28002         -48.99%
BenchmarkInsertChain_valueTx_memdb-8            100512        48822         -51.43%
BenchmarkInsertChain_valueTx_diskdb-8           115107        62805         -45.44%
BenchmarkInsertChain_valueTx_100kB_memdb-8      1376513       899592        -34.65%
BenchmarkInsertChain_valueTx_100kB_diskdb-8     3008558       2297909       -23.62%
BenchmarkInsertChain_uncles_memdb-8             68362         41908         -38.70%
BenchmarkInsertChain_uncles_diskdb-8            79106         55894         -29.34%
BenchmarkInsertChain_ring200_memdb-8            94582187      18819301      -80.10%
BenchmarkInsertChain_ring200_diskdb-8           96704176      20032446      -79.28%
BenchmarkInsertChain_ring1000_memdb-8           137721565     23260916      -83.11%
BenchmarkInsertChain_ring1000_diskdb-8          139818770     24621324      -82.39%
BenchmarkCompactEncode                          16            4             -75.00%
BenchmarkCompactDecode                          16            16            +0.00%
BenchmarkCompactHexDecode                       32            32            +0.00%
BenchmarkDecodeCompact                          4             4             +0.00%
BenchmarkGet                                    144           80            -44.44%
BenchmarkGetDB                                  144           320           +122.22%
BenchmarkUpdateBE                               1766          1567          -11.27%
BenchmarkUpdateLE                               1797          1627          -9.46%
BenchmarkHashBE                                 223376483     3735414       -98.33%
BenchmarkHashLE                                 244950514     3927063       -98.40%
```