@zelig 2015-08-20
is ethdb.Database put/get meant to be concurrently used? cos AFAIU LDBDatabase implementation is threadsafe due to leveldb and batch calls being safe but MemDatabase is not, nor is its batch.