@fjl 2015-08-20
Trie is not thread safe either.