@zelig 2015-08-20
Yes the trie is clearly documented not to be. But ethdb.Database is independent and ldbd is actually perfectly functional concurrently right? Either way it's worth a comment