@fjl 2015-09-09
I've rebased this and added my initial stab at the merkle proof functions. Some comments on those below. My plan is to move parts of this (rlp changes #1778, ethdb changes) into develop soon-ish and wait for the other pending core changes to be merged before bringing in the trie. This also leaves more time for reviewing.

About the merkle proof code: @ebuchman did an implementation for the old trie in #1624, but I couldn't port these directly because the API wasn't quite right. For the light client, the initial use of these proofs is that we have the state root hash and wish to retrieve the proof and value for a key in one request. The API in #1624 expected the value to be passed as an input to `Verify`. In addition, the PR defined a custom encoding for proofs. AFAIK the plan is to represent proofs [as a list of encoded nodes as defined in the YP](https://github.com/ethereum/wiki/wiki/Ethereum-Wire-Protocol#proposal-pv64).

@zsfelfoldi Please comment on the API. The two functions of interest are

```
func (*Trie) Prove(key []byte) []rlp.RawValue
func VerifyProof(root common.Hash, key []byte, proof []rlp.RawValue) (value []byte, err error)
```